refactor: co-locate component ownership
