+++ src/app/provider.tsx
@@ -8,8 +8,8 @@ import { AISearch } from '@/components/features/assistant/chat'
 import { AISearchTrigger } from '@/components/features/assistant/search-trigger'
 import { MobileNav } from '@/components/layout/header/mobile/mobile-nav'
 import Analytics from '@/components/providers/analytics'
-import { SmoothCursor } from '@/components/shared/smooth-cursor'
-import { TailwindIndicator } from '@/components/shared/tailwind-indicator'
+import { SmoothCursor } from '@/components/providers/smooth-cursor'
+import { TailwindIndicator } from '@/components/providers/tailwind-indicator'
 import { Toaster } from '@/components/ui/sonner'
 import { TooltipProvider } from '@/components/ui/tooltip'
