+++ src/components/features/assistant/copy-button.tsx
