--- src/components/icons/animated/upload.tsx
@@ -1,72 +1,0 @@
-'use client'
-
-import type { Variants } from 'motion/react'
-import { motion, useAnimation } from 'motion/react'
-import type { Ref, SVGProps } from 'react'
-import { useImperativeHandle } from 'react'
-
-import { cn } from '@/lib/utils'
-
-export interface UploadIconHandle {
-  startAnimation: () => void
-  stopAnimation: () => void
-}
-
-interface UploadIconProps extends Omit<SVGProps<SVGSVGElement>, 'ref'> {
-  ref?: Ref<UploadIconHandle>
-  size?: number
-}
-
-const ARROW_VARIANTS: Variants = {
-  animate: {
-    transition: {
-      damping: 10,
-      mass: 1,
-      stiffness: 200,
-      type: 'spring',
-    },
-    y: -2,
-  },
-  normal: { y: 0 },
-}
-
-const UploadIcon = ({
-  className,
-  size = 28,
-  ref,
-  ...props
-}: UploadIconProps) => {
-  const controls = useAnimation()
-  useImperativeHandle(ref, () => ({
-    startAnimation: () => controls.start('animate'),
-    stopAnimation: () => controls.start('normal'),
-  }))
-
-  return (
-    <svg
-      aria-hidden={props['aria-label'] ? undefined : true}
-      className={cn(className)}
-      fill='none'
-      height={size}
-      role={props['aria-label'] ? 'img' : undefined}
-      stroke='currentColor'
-      strokeLinecap='round'
-      strokeLinejoin='round'
-      strokeWidth='2'
-      viewBox='0 0 24 24'
-      width={size}
-      xmlns='http://www.w3.org/2000/svg'
-      {...props}
-    >
-      <path d='M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4' />
-      <motion.g animate={controls} variants={ARROW_VARIANTS}>
-        <polyline points='17 8 12 3 7 8' />
-        <line x1='12' x2='12' y1='3' y2='15' />
-      </motion.g>
-    </svg>
-  )
-}
-
-UploadIcon.displayName = 'UploadIcon'
-
-export { UploadIcon }
